Do not use DocumentBuilder.parse to read settings
DocumentBuilder.parse was attempting to load the xml schema from its
hosting location at Oracle. This was resulting in a "too many redirects"
error because Oracle's hosting was broken. I replaced DocumentBuilder.parse,
with the Java properties class which works with this XML schema by default
and does not attemt to remtely fetch the xml schema.